Move merkle state tests to integration path
Move the merkle state tests to the integration path (under
libtransact/tests - this is the cargo default path).  These tests can
then be reapplied for the different database types.
